Move tokens to own module
